refactor: move storage and utility files out of lib/others (#1898)
