+++ lib/storage/badge_text_storage.dart
