+++ lib/utils/badge_utils.dart
