+++ lib/utils/qr_code_helper.dart
