+++ lib/view/qr_scan_screen.dart
@@ -1,5 +1,5 @@
 import 'package:badgemagic/constants.dart';
-import 'package:badgemagic/others/qr_code_helper.dart';
+import 'package:badgemagic/utils/qr_code_helper.dart';
 import 'package:badgemagic/others/toast_utils.dart';
 import 'package:badgemagic/l10n/app_localizations.dart';
 import 'package:badgemagic/others/localization_service.dart';
